# Avoid extra allocations in calls to Reset()