During an investigation of performance of onne of our services, we noticed a large amount of CPU spent on calls to `Reset`. Looking at the code, we determined that the `Reset` function allocates new memory for each of the registers every time it's called. This PR aims to change that behavior by simply zeroing out each register instead of reallocating them