@joepin 2020-03-02
Added a [benchmark test for this change](https://github.com/DataDog/hyperloglog/pull/12/commits/453e9227a1c953b9318470a7a2e3437edda598e8), and ran it with `go test -bench=BenchmarkReset -benchmem` on both the old code and new code. Using [benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp), here are the results:

```
$ benchcmp old.txt new.txt 
benchmark            old ns/op     new ns/op     delta
BenchmarkReset-8     15207         15611         +2.66%

benchmark            old allocs     new allocs     delta
BenchmarkReset-8     1              0              -100.00%

benchmark            old bytes     new bytes     delta
BenchmarkReset-8     256           0             -100.00%
```